* main: (48 commits)
  Update release notes.
  Refactor macro _FillValue to NC_FillValue in support of https://github.com/Unidata/netcdf-c/issues/2858
  Remove appveyor config file.
  Add CI for a Windows Runner on Github Actions.
  Add cmake prefix path to appveyor config.
  Attempt to fix zlib-related error in appveyor.
  Correct typo.
  Remove check for libcurl unless it is necessary for required functionality.
  Rename the vendored strlcat symbol
  Fix conversion warnings in libdispatch
  Set flags to avoid warning messages if curl isn't found.
  Use modern cmake nomenclature for curl.
  Fix truncated-format warning in ncgen
  Fix some conversion warnings in ncgen3
  Fix some conversion warnings in ncgen3 generated files
  Regenerate ncgen3
  CMake: Add option to automatically regenerate ncgen3/ncgen
  Skip checking for duplicates if only one element in list
  Change format of backwards-loops
  Silence some conversion warnings in ncgen generated files
  ...

+++ ncgen/data.h
@@ -3,6 +3,7 @@
  *   See netcdf/COPYRIGHT file for copying and redistribution conditions.
  *********************************************************************/
 
+#include <stddef.h>
 #ifndef DATA_H
 #define DATA_H 1
 
@@ -36,11 +37,11 @@ typedef union Constvalue {
     float floatv;               /* NC_FLOAT*/
     double doublev;             /* NC_DOUBLE*/
     struct Stringv {		/* NC_STRING*/
-	int len;
-	char* stringv; 
+        size_t len;
+        char* stringv;
     } stringv;
     struct Opaquev {     /* NC_OPAQUE*/
-	int len; /* length as originally written (rounded to even number)*/
+	size_t len; /* length as originally written (rounded to even number)*/
 	char* stringv; /*as  constant was written*/
 		      /* (padded to even # chars >= 16)*/
 		      /* without leading 0x*/
